Refatoraçao geral
