+++ include/DMASerial.h
@@ -2,9 +2,15 @@
 #define DMASERIAL_H_
 
 #include <Arduino.h>
-#include <DMAInterrupts.h>
 #include <libmaple/dma.h>
 #include <libmaple/usart.h>
+
+typedef enum {
+  DMA_IRQ_HANDLER_NONE = 0,
+  DMA_IRQ_HANDLER_1 = 1,
+  DMA_IRQ_HANDLER_2 = 2,
+  DMA_IRQ_HANDLER_3 = 3
+} dma_irq_handler;
 
 class DMASerial {
  public:
@@ -36,6 +42,24 @@ class DMASerial {
   uint8_t* buffer;
   uint8_t index;
   bool transfering;
+};
+
+class DMAInterrupts {
+ public:
+  static DMASerial* serial1;
+  static DMASerial* serial2;
+  static DMASerial* serial3;
+
+  static void set_interrupt_handler(dma_irq_handler, DMASerial*);
+  static void remove_interrupt_handler(dma_irq_handler);
+
+  static void handler1(void);
+  static void handler2(void);
+  static void handler3(void);
+
+ private:
+  DMAInterrupts();
+  ~DMAInterrupts();
 };
 
 #endif
